refactor(computer-use): share structured_content() between app.py and preflight.py
#303's embedded-host permissions check (preflight.py::_embedded_permissions)
re-derived app.py::structured_content()'s exact key-casing-tolerant extraction
inline instead of reusing it. Moved the helper to result.py (already the
SDK-free home shared across the runner/supervisor/cli boundary) so preflight.py
can call it without importing the SDK-dependent app.py; app.py and cli.py now
import it from there too.

No behavior change: same two-key fallback and dict-type check, just named
once. Verified via the full test suite (497 passed / 1 skipped, matching the
pre-change baseline exactly) and `ruff check .` clean.
